Deleted TabBar component and add NavigationContainer inside index.ts
